GraphRAG Pipeline/requirements.txt
google-genai>=1.0.0
faiss-cpu>=1.7.4
networkx>=3.0
httpx>=0.25.0
numpy>=1.24.0
pytest>=8.0.0
